+++ src/main/kotlin/com/tribus/markdown/editor/MarkdownFileEditorListener.kt
@@ -107,6 +107,12 @@ class MarkdownFileEditorListener : EditorFactoryListener {
                 "com.tribus.markdown.actions.ToggleMath" to CustomShortcutSet(
                     KeyboardShortcut(KeyStroke.getKeyStroke(KeyEvent.VK_M, InputEvent.CTRL_DOWN_MASK), null)
                 ),
+                "com.tribus.markdown.actions.InsertLink" to CustomShortcutSet(
+                    KeyboardShortcut(KeyStroke.getKeyStroke(KeyEvent.VK_K, InputEvent.CTRL_DOWN_MASK), null)
+                ),
+                "com.tribus.markdown.actions.InsertImage" to CustomShortcutSet(
+                    KeyboardShortcut(KeyStroke.getKeyStroke(KeyEvent.VK_K, InputEvent.CTRL_DOWN_MASK or InputEvent.SHIFT_DOWN_MASK), null)
+                ),
                 "com.tribus.markdown.actions.ToggleBlockquote" to CustomShortcutSet(
                     KeyboardShortcut(KeyStroke.getKeyStroke(KeyEvent.VK_PERIOD, InputEvent.CTRL_DOWN_MASK or InputEvent.SHIFT_DOWN_MASK), null)
                 ),